Fix issues preventing compilation on macOS
We need to bump our deployment target to 10.15 in order to be able to use std::filesystem::path. This should not be a problem, since that's the official minspec version.

Sinplify the definition of a couple of macros, making them into CMake defines rather than creating a file. The file was not being picked up by the macOS build.

Add definitions for exporting the API when not compiling under Windows.

Ensure consistent types for variables.

+++ src/ResourceGroupImpl.cpp
@@ -995,7 +995,7 @@ namespace CarbonResources
 			ResourceInfo* resourcePrevious = resourceGroupSubtractionPrevious.m_resourcesParameter.At( i );
 
 			ResourceInfo* resourceNext = resourceGroupSubtractionLatest.m_resourcesParameter.At( i );
-        	uint64_t patchSourceOffset{0};
+        	size_t patchSourceOffset{0};
         	uint64_t patchSourceOffsetDelta{0};
 
             // Check to see if previous entry contains dummy information
@@ -1103,7 +1103,7 @@ namespace CarbonResources
 					std::string patchData;
 
 					bool chunkMatchFound{false};
-					uint64_t matchCount{0};
+					size_t matchCount{0};
 
 
                     if (previousFileData != "")
